Update Homepage UI
Explanation:
The given code diff adds a new file 'page.jsx' with the Homepage UI implementation. Thus, the commit message needs to be something that clearly expresses this change. Here, the message 'Update Homepage UI' is a concise way of making it clear that the homepage UI has been implemented/updated. This message follows the present tense format recommended by Git for commit messages i.e., it states that the action of updating the homepage UI has been performed.

Final answer:
Update Homepage UI

Note: Just the commit message is given as an answer as specified.