"Update work page with new project data and Swiper component"
The commit message "Update work page with new project data and Swiper component" is a concise and clear summary of the changes made in the code diff. It is written in present tense and is a maximum of 50 characters, as specified. The message conveys that the work page has been updated with new project data and a Swiper component has been added, which are the main modifications in the provided code diff.